[14.0][MIG] - stock_available_unreserved
