+++ stock_available_unreserved/models/__init__.py
@@ -1,2 +1,3 @@
-from . import product
-from . import quant
+from . import product_template
+from . import product_product
+from . import stock_quant
