+++ stock_available_unreserved/readme/CONTRIBUTORS.rst
@@ -3,3 +3,4 @@
 * Mykhailo Panarin <m.panarin@mobilunity.com>
 * Atte Isopuro <atte.isopuro@avoin.systems>
 * Lois Rilo <lois.rilo@forgeflow.com>
+* Souheil Bejaoui <souheil.bejaoui@acsone.eu>
